% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_GLMs.R
\name{calibration_GLMs}
\alias{calibration_GLMs}
\title{GLM model calibration}
\usage{
calibration_GLMs(
  dependent,
  data,
  cv_kfolds = 5,
  cformulas = NULL,
  type = "l",
  parallel = FALSE,
  ncores = 4,
  seed = 1,
  weights = NULL
)
}
\arguments{
\item{dependent}{\code{character}, name of dependent variable.}

\item{data}{data.frame or matrix of independent variables.}

\item{cv_kfolds}{\code{numeric}, number of folds to be using  k-fold
cross-validation.}

\item{cformulas}{a vector of character with the set of formulas to test.
Default=NULL.}

\item{type}{\code{character}, a character string that must contain "l", "p", "q"
or a combination of them. l = lineal, q = quadratic,
p = interaction between two variables. Default = "l".}

\item{parallel}{\code{logical}, choose to run on parallel or sequential.
Default = NULL.}

\item{ncores}{\code{numeric}, number of cores to use. Default = 4.}

\item{seed}{\code{numeric}, integer value to specify an initial seed. Default = 1.}

\item{weights}{a vector with the weights for observations.}
}
\value{
a list of dataframe.
}
\description{
GLM model calibration
}
