% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{enm_data}
\alias{enm_data}
\title{Example data used to run model calibration exercises}
\format{
A data frame with 5627 rows and 3 columns.
\describe{
\item{Sp}{numeric, values of 0 = absence and 1 = presence.}
\item{bio_1}{numeric, temperature values.}
\item{bio_12}{numeric, precipitation values.}
}
}
\usage{
enm_data
}
\description{
A dataset containing information on presence and absence, and
independent variables used to fit GLM models.
}
\examples{
data("enm_data", package = "enmpa")
head(enm_data)
}
\keyword{datasets}
