% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{enmpa_calibration}
\alias{enmpa_calibration}
\alias{new_enmpa_calibration}
\title{Constructor of S3 objects of class enmpa_calibration}
\usage{
new_enmpa_calibration(selected, summary, calibration_results, data,
                      partitioned_data, weights = NULL)
}
\arguments{
\item{selected}{date.frame with information about selected models.}

\item{summary}{data.frame a summary of statistics for all models.}

\item{calibration_results}{data.frame with results obtained from
cross-validation for all models.}

\item{data}{data.frame or matrix with the input data used for calibration.}

\item{partitioned_data}{a list of partition indices.}

\item{weights}{(numeric) a vector with the weights for observations.
Default = NULL.}
}
\value{
An S3 object of class \code{enmpa_calibration}.
}
\description{
Constructor of S3 objects of class enmpa_calibration
}
