% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Fitting the selected models}
\usage{
fit_glm(formulas, data, weights = NULL)
}
\arguments{
\item{formulas}{(character) a vector containing the formula(s) of model(s).}

\item{data}{data.frame or matrix with the dependent and independent variables.}

\item{weights}{(numeric) a vector with the weights for observations.}
}
\value{
a list of fitted models
}
\description{
Function that facilitates the fitting step of multiple models.
}
