% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_evaluation.R
\name{indep_evaluation}
\alias{indep_evaluation}
\title{Evaluate final models using an independent dataset.}
\usage{
indep_evaluation(indep_data, prediction, crs, occ = "sp", xy = c("lon", "lat"))
}
\arguments{
\item{indep_data}{\code{data.frame}.}

\item{prediction}{\code{SpatRaster} object.}

\item{crs}{\code{character}, any Coordinate Reference System (CRS)
that terra::vect() function accepts.}

\item{occ}{\code{character}, column name with the occurrence data.}

\item{xy}{\code{character} vector, field names associated with the geometry data,
representing the x and y coordinates.}
}
\value{
\code{data.frame}
}
\description{
The function enables a final evaluation of model predictions using a dataset
that is independent of the one used for model calibration.
}
