% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_partition.R
\name{kfold_partition}
\alias{kfold_partition}
\title{K-fold data partitioning}
\usage{
kfold_partition(data, k = 2, seed = 1)
}
\arguments{
\item{data}{data.frame or matrix containing at least two columns. The first
column must contain absence and presence records as 0 and 1.}

\item{k}{\code{numeric}, the number of groups that the given data is to be split
into.}

\item{seed}{\code{numeric}, integer value to specify an initial seed. Default = 1.}
}
\value{
A list of vectors with the indices of rows corresponding to each fold.
}
\description{
K-fold data partitioning
}
