% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{model_validation}
\alias{model_validation}
\title{Model validation options}
\usage{
model_validation(formula, data, family = binomial(link = "logit"),
                 weights = NULL, cv = FALSE, partition_index = NULL,
                 k = NULL, n_threshold = 100, seed = 1)
}
\arguments{
\item{formula}{a formula.}

\item{data}{data set to create model}

\item{family}{family of model to be used}

\item{weights}{weights for observations}

\item{cv}{whether to use a cross-validation approach}

\item{partition_index}{list of indices for cross-validation in k-fold}

\item{k}{number of folds for a new k-fold index preparation. Ignored if
\code{partition_index} is defined.}

\item{n_threshold}{\code{numeric}number of threshold values to be used.
Default = 100.}

\item{seed}{seed.}
}
\value{
data.frame
}
\description{
Model evaluation using entire set of data or a cross-validation approach.
}
