% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot variable importance}
\usage{
plot_importance(x,
                xlab = NULL,
                ylab = "Relative contribution",
                main = "Variable importance",
                extra_info = TRUE, ...)
}
\arguments{
\item{x}{data.frame output from \code{\link{var_importance}}.}

\item{xlab}{(character) a label for the x axis.}

\item{ylab}{(character) a label for the y axis.}

\item{main}{(character) main title for the plot.}

\item{extra_info}{(logical) adds extra details to the plot when there is
more information from more than one model.}

\item{...}{additional arguments passed to \code{\link[graphics]{barplot}} or
\code{\link[graphics]{boxplot}}.}
}
\description{
It facilitates the visualization of the results obtained by the function
\code{\link{var_importance}}.
}
\examples{
# Load species occurrences and environmental data.
enm_data <- read.csv(system.file("extdata", "pa_data.csv", package = "enmpa"))

# Custom formulas
forms <- c("Sp ~ bio_1 + I(bio_1^2) + I(bio_12^2)",
           "Sp ~ bio_12 + I(bio_1^2) + I(bio_12^2)")

# Fitting models
fits <- fit_glm(forms, data = enm_data )

# Variable importance for single models
vi_1 <- var_importance(fits$Model_ID_1)
plot_importance(x = vi_1)

vi_2 <- var_importance(fits$Model_ID_2)
plot_importance(x = vi_2)

# Variable importance for multiple models
vi_c <- var_importance(fits)
plot_importance(x = vi_c)

}
