% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_selected.R
\name{predict_selected}
\alias{predict_selected}
\title{Predictions for the models selected after calibration}
\usage{
predict_selected(glm_calibration, newdata, clamping = FALSE,
                 type = "response", consensus = TRUE)
}
\arguments{
\item{glm_calibration}{a list resulting from \code{\link{calibration_glm}}.
Predictions are from models in the slot "selected".}

\item{newdata}{a \code{SpatRaster}, data.frame, or matrix with the new data on
which to predict.}

\item{clamping}{(logical) this option controls extrapolation when making
predictions for environmental conditions beyond the calibration data.
Default = FALSE.}

\item{type}{(character) the type of prediction required. For a default
binomial model the default predictions are of log-odds (probabilities on
logit scale). The default, "response", returns predicted probabilities.}

\item{consensus}{(logical) valid if \code{newdata} is a \code{SpatRaster}, whether to
produce consensus results obtained by combining the predictions from the
collection of selected models. By default consensuses are calculated using
the mean, median, a weighted average using the AIC weights, and variance.
Default = TRUE.}
}
\value{
A list with predictions of selected models on the \code{newdata}. Consensus
predictions are added if defined and if \code{newdata} is a \code{SpatRaster} object.
}
\description{
Wrapper function that facilitates the prediction of those models
selected as the most robust. In addition, it allows the calculation of
consensus models, when more than one model are selected.
}
\examples{
# Load two fitted models
data(cal_res, package = "enmpa")

# Load raster layers to be projected
env_vars <- terra::rast(system.file("extdata", "vars.tif", package = "enmpa"))

# Predictions (only one selected mode, no consensus required)
preds <- predict_selected(cal_res, newdata = env_vars, consensus = FALSE)

# Plot prediction
terra::plot(preds$predictions)
}
