% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print}
\alias{print}
\alias{print.enmpa_calibration}
\alias{print,enmpa_calibration-method}
\alias{print,enmpa_fitted_models-method}
\alias{print.enmpa_fitted_models}
\title{Print a short version of elements in 'calibration' and 'fitted models' objects}
\usage{
\method{print}{enmpa_calibration}(x, ...)

\method{print}{enmpa_fitted_models}(x, ...)
}
\arguments{
\item{x}{object of enmpa_fitted_models or enmpa_calibration}

\item{...}{additional arguments affecting the summary produced. Ignored in
these functions.}
}
\description{
Print a short version of elements in 'calibration' and 'fitted models' objects
}
