% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sel_fit}
\alias{sel_fit}
\title{Example of selected models fitted}
\format{
A list with four elements.
\describe{
\item{glms_fitted}{a fitted glm (ModelID_7).}
\item{selected}{a data.frame with results from evaluation of ModelID_7}
\item{data}{a data.frame containing information on presence and absence, and
independent variables used to fit GLM models.}
\item{weights}{a vector with the weights for observations.}
}
}
\usage{
sel_fit
}
\description{
An object of the class enmpa_fitted_models containing
fitted selected model(s) and the information from model evaluation for such
model(s).
}
\examples{
data("sel_fit", package = "enmpa")
}
\keyword{datasets}
