% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary}
\alias{summary}
\alias{summary.enmpa_calibration}
\alias{summary,enmpa_calibration-method}
\alias{summary,enmpa_fitted_models-method}
\alias{summary.enmpa_fitted_models}
\title{Summary of 'calibration' and 'fitted models'}
\usage{
\method{summary}{enmpa_calibration}(object, ...)

\method{summary}{enmpa_fitted_models}(object, ...)
}
\arguments{
\item{object}{of class enmpa_calibration or enmpa_fitted_models}

\item{...}{additional arguments affecting the summary produced. Ignored in
these functions.}
}
\value{
A printed summary.
}
\description{
Summary of 'calibration' and 'fitted models'
}
