% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{test}
\alias{test}
\title{Example data used to test models}
\format{
A data frame with 100 rows and 3 columns.
\describe{
\item{Sp}{numeric, values of 0 = absence and 1 = presence.}
\item{lon}{numeric, longitude values.}
\item{lat}{numeric, latitude values.}
}
}
\usage{
test
}
\description{
A dataset containing information on presence and absence, and
independent variables used to fit GLM models.
}
\examples{
data("test", package = "enmpa")
head(test)
}
\keyword{datasets}
