% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cal_res}
\alias{cal_res}
\title{Example of results obtained from GLM calibration using enmpa}
\format{
An object of class enmpa_calibration with results from the function
\verb{\link{calibration_glm}}.
}
\usage{
cal_res
}
\description{
An object of the class enmpa_calibration storing the results from GLM
calibration.
}
\examples{
data("cal_res", package = "enmpa")

str(cal_res)
}
\keyword{datasets}
