% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmpa.R
\docType{package}
\name{enmpa}
\alias{enmpa}
\title{enmpa: Ecological Niche Modeling using Presence-Absence Data}
\description{
\code{enmpa} contains a set of tools to perform Ecological Niche Modeling with
presence-abscense data.
}
\details{
It includes algorithms for data partitioning, model fitting, calibration,
evaluation, selection, and prediction. Other functios help to explore model
features as such variable response curves and variable importance.
}
\section{Functions in enmpa}{

\code{\link{calibration_glm}}, \code{\link{get_formulas}},
\code{\link{indep_evaluation}}, \code{\link{kfold_partition}}, add others
}

\section{Data included}{

add data names as with the functions here
}

