% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmpa.R
\docType{package}
\name{enmpa}
\alias{enmpa}
\title{enmpa: Ecological Niche Modeling using Presence-Absence Data}
\description{
\code{enmpa} contains a set of tools to perform Ecological Niche Modeling with
presence-absence data.
}
\details{
It includes algorithms for data partitioning, model fitting, calibration,
evaluation, selection, and prediction. Other functions help to explore model
features as such variable response curves and variable importance.
}
\section{Main functions in \code{enmpa}}{

\code{\link{calibration_glm}}, \code{\link{get_formulas}},
\code{\link{predict_selected}}, \code{\link{response_curve}},
\code{\link{var_importance}}, \code{\link{independent_eval}}
add others
}

\section{Data included}{

add data names as with the functions here
}

