% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmpa.R
\docType{package}
\name{enmpa}
\alias{enmpa}
\title{enmpa: Ecological Niche Modeling using Presence-Absence Data}
\description{
\code{enmpa} contains a set of tools to perform detailed Ecological Niche Modeling
using presence-absence data.
}
\details{
It includes algorithms for data partitioning, model fitting, calibration,
evaluation, selection, and prediction. Other functions help to explore model
features as such variable response curves and variable importance.
}
\section{Main functions in \code{enmpa}}{

\code{\link{calibration_glm}}, \code{\link{evaluation_stats}},
\code{\link{fit_glms}}, \code{\link{fit_selected}},
\code{\link{get_formulas}}, \code{\link{independent_eval1}},
\code{\link{kfold_partition}}, \code{\link{model_selection}}
\code{\link{model_validation}}, \code{\link{niche_signal}},
\code{\link{optimize_metrics}}, \code{\link{predict_glm}},
\code{\link{predict_selected}}, \code{\link{response_curve}},
\code{\link{var_importance}}
}

