% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formulas.R
\name{get_formulas}
\alias{get_formulas}
\title{Get formulas according to types of responses needed}
\usage{
get_formulas(dependent, independent, type = "l", all_combinations = TRUE,
             minvar=1, maxvar = NULL)
}
\arguments{
\item{dependent}{(character) name of dependent variable.}

\item{independent}{(character) a vector of names of independent variables.}

\item{type}{(character) a character string that must contain "l", "p", "q"
or a combination of them. l = lineal, q = quadratic,
p = interaction between two variables. Default = "l".}

\item{all_combinations}{(logical) whether to produce all combinations,
default = TRUE. FALSE returns only the most complex formula defined in type.}

\item{minvar}{(numeric) minimum number of features.}

\item{maxvar}{(numeric) maximum number of features.}
}
\value{
A character vector containing the resulting formula(s).
}
\description{
Generate forms based on independent variables that can predict the dependent
variable, taking into account the required response types. If necessary,
calculate all possible combinations of these formulas considering different
variable combinations.
}
\examples{
# example variables
dep <- "presence"
ind <- c("temperature", "humidity")

# one formula according to "type"
formula <- get_formulas(dep, ind, type = "lqp", all_combinations = FALSE)

# all combinations according to type
formulas <- get_formulas(dep, ind, type = "lqp", all_combinations = TRUE)
}
