% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formulas.R
\name{get_formulas}
\alias{get_formulas}
\title{Get formulas according to types of responses needed}
\usage{
get_formulas(dependent, independent, type = "l", all_combinations = TRUE,
minvar=1, maxvar = NULL)
}
\arguments{
\item{dependent}{\code{character}, name of dependent variable.}

\item{independent}{\code{character}, a vector of names of independent variables.}

\item{type}{\code{character}, a character string that must contain "l", "p", "q"
or a combination of them. l = lineal, q = quadratic,
p = interaction between two variables. Default = "l".}

\item{all_combinations}{\code{logical}, whether to produce all combinations,
default = TRUE. FALSE returns only the most complex formula defined in type.}

\item{minvar}{\code{numeric} minimum number of features.}

\item{maxvar}{\code{numeric} maximum number of features.}
}
\value{
A character vector containing the resulting formula(s).
}
\description{
Produces forms in which independent variables can predict the dependent one,
considering response types required. All possible combinations of these
formulas considering combinations of variables can be calculated if needed.
}
\examples{
# example variables
dep <- "presence"
ind <- c("temperature", "humidity")

# one formula according to "type"
formula <- get_formulas(dep, ind, type = "lqp", all_combinations = FALSE)

# all combinations according to type
formulas <- get_formulas(dep, ind, type = "lqp", all_combinations = TRUE)
}
