% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formulas.R
\name{get_formulas}
\alias{get_formulas}
\alias{get_formulas_main}
\alias{aux_var_comb}
\alias{aux_string_comb}
\title{Get GLM formulas according to types of responses needed}
\usage{
get_formulas(dependent, independent, type = "l", mode = "moderate",
             minvar = 1, maxvar = NULL)

get_formulas_main(dependent, independent, type = "l",
                  complex = FALSE, minvar = 1, maxvar = NULL)

aux_var_comb(var_names, minvar = 2, maxvar = NULL)

aux_string_comb(string)
}
\arguments{
\item{dependent}{(character) name of dependent variable.}

\item{independent}{(character) a vector of names of independent variables.}

\item{type}{(character) a character string that must contain the letters "l",
"p", "q", or a combination of them. l = lineal, q = quadratic,
p = interaction between two variables. Default = "l".}

\item{mode}{(character) a character string to indicate the strategy to create
formulas, must be among "light", "moderate", "intensive" or "complex".
Default = "moderate".}

\item{minvar}{(numeric) minimum number of independent variables in formulas.}

\item{maxvar}{(numeric) maximum number of independent variables in formulas.}

\item{complex}{(logical) whether to return the most complex formula.}

\item{var_names}{sames as \code{dependent}.}

\item{string}{same as \code{type}.}
}
\value{
A character vector containing the resulting formula(s).
}
\description{
Generate GLM formulas for independent variables predicting a dependent
variable, taking into account response types required. All possible
combinations of variables can be created using arguments of the function
}
\details{
\code{mode} options determine the strategy to combine independent variables
depending on \code{type}:
\itemize{
\item \strong{light} - returns simple iterations of complex formulas.
\item \strong{moderate} - returns a vector of formulas with increasing number of
independent variables.
\item \strong{intensive} - returns all possible combinations of variables. This can be
time-consuming for 6 or more independent variables.
\item \strong{complex} - returns only the most complex formula.
}
}
\examples{
# example variables
dep <- "sp"
ind <- c("temp", "rain", "slope")

# The most complex formula according to "type"
get_formulas(dep, ind, type = "lqp", mode = "complex")

# mode = 'light', combinations according to type
get_formulas(dep, ind, type = "lqp", mode = "light")

# mode = 'light', combinations according to type
get_formulas(dep, ind, type = "lqp", mode = "intensive")
}
