% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formulas.R
\name{get_formulas}
\alias{get_formulas}
\title{Get formula combinations}
\usage{
get_formulas("species", data = data, type = "l")
}
\arguments{
\item{resp_var}{\code{character}, name of the column with the presence-absence data.}

\item{data}{a data.frame or  matrix of variables}

\item{type}{\code{character}, a character string that must contain 'l', 'p', 'q'
or a combination of them. l= lineal q = quadratic p = interaction between two variables.
Default = 'l'.}

\item{all_comb}{\code{logical}, return all combination (Default=TRUE) or only the most complex formula (FALSE).}
}
\value{
a character vector containing the resulting formula combination(s).
}
\description{
It provides a method for calculating all possible or the most complex formula(s).
}
