% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_partition.R
\name{kfold_partition}
\alias{kfold_partition}
\title{K-fold data partitioning}
\usage{
kfold_partition(data, occ, k = 2, seed = 1)
}
\arguments{
\item{data}{data.frame or matrix containing at least two columns.}

\item{occ}{Column name that contains the absence and presence records as 0
and 1.}

\item{k}{(numeric) the number of groups that the given data is to be split
into.}

\item{seed}{(numeric) integer value to specify an initial seed. Default = 1.}
}
\value{
A list of vectors with the indices of rows corresponding to each fold.
}
\description{
The function is designed to partition the available data into k equal-sized
subsets or folds while maintaining the global proportion of presence-absences
in each fold. This approach leads to a more robust evaluation of the model's
performance, as it accounts for the inherent variability in the data.
}
\examples{

data <- data.frame(species = c(rep(0, 80), rep (1,20)),
                   variable1 = rnorm(100),
                   variable2 = rpois(100, 2))

kfolds <- kfold_partition(data, occ = "species", k = 2)

data[kfolds$Fold_1,]
data[kfolds$Fold_2,]


}
