% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{model_validation}
\alias{model_validation}
\title{Model validation options}
\usage{
model_validation(formula, data, family = binomial(link = "logit"),
                 weights = NULL, cv = FALSE, partition_index = NULL,
                 k = NULL, dependent = NULL, n_threshold = 100,
                 keep_coefficients = FALSE, seed = 1)
}
\arguments{
\item{formula}{(character) \code{expressions} to return a \code{formula} object class.}

\item{data}{data.frame to create model.}

\item{family}{a \code{family} object for models used by functions such as \code{glm}.}

\item{weights}{(numeric) vector with weights for observations.}

\item{cv}{(logical) whether to use a k-fold cross validation.}

\item{partition_index}{list of indices for cross validation in k-fold.}

\item{k}{(numeric) number of folds for a new k-fold index preparation.
Ignored if \code{partition_index} is defined.}

\item{dependent}{(character) name of dependent variable. Ignore if
\code{cv} = FALSE.}

\item{n_threshold}{(numeric) number of threshold values to be used.
Default = 100.}

\item{keep_coefficients}{(logical) whether to keep model coefficients.
Default = FALSE.}

\item{seed}{(numeric) a seed number.}
}
\value{
data.frame
}
\description{
Model evaluation using entire set of data or a K-fold cross validation
approach. Models are assessed based on discrimination power (ROC-AUC),
classification capability (FPR, accuracy, sensitivity, specificity, and TSS ),
and the balance between goodness-of-fit and complexity (number of parameters
and AICc).
}
\examples{
# Load species occurrences and environmental data.
enm_data <- read.csv(system.file("extdata", "pa_data.csv", package = "enmpa"))
head(enm_data)

# Custom formula
form <- c("Sp ~ bio_1 + I(bio_1^2) + I(bio_12^2)")

# Model evaluation using the entire set of records
model_validation(form, data = enm_data)

# Model evaluation using a k-fold cross-validation (k = 3)
model_validation(form, data = enm_data, cv = TRUE, k = 3, dependent = "Sp")


}
