% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{model_validation}
\alias{model_validation}
\title{Model validation options}
\usage{
model_validation(formula, data, family = binomial(link = "logit"),
                 weights = NULL, cv = FALSE, partition_index = NULL,
                 k = NULL, n_threshold = 100, keep_coefficients = FALSE,
                 seed = 1)
}
\arguments{
\item{formula}{\code{character}, “expressions” to return a "formula" object class.}

\item{data}{\code{data.frame}, data set to create model}

\item{family, }{a family object for models used by functions such as \code{glm}.}

\item{weights}{\code{numeric} vector with weights for observations.}

\item{cv}{\code{logical}, whether to use a cross-validation apprach}

\item{partition_index}{\code{list} of indices for cross-validation in k-fold.}

\item{k}{\code{numeric}, number of folds for a new k-fold index preparation. Ignored if
\code{partition_index} is defined.}

\item{n_threshold}{\code{numeric}, number of threshold values to be used.
Default = 100.}

\item{keep_coefficients}{\code{logical}, whether to keep model coefficients.
Default = FALSE.}

\item{seed}{\code{numeric}, seed.}
}
\value{
data.frame
}
\description{
Model evaluation using entire set of data or a cross-validation approach.
}
