% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot variable importance}
\usage{
plot_importance(
  x,
  xlab = NULL,
  ylab = "Relative contribution",
  main = "Variable importance",
  extra_info = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame output from \code{\link{enmpa::var_importance}}.}

\item{xlab}{(character) a label for the x axis.}

\item{ylab}{(character) a label for the y axis.}

\item{main}{(character) main title for the plot.}

\item{extra_info}{(logical) adds extra details to the plot when there is
more information from more than one model.}

\item{...}{additional arguments passed to \code{\link[graphics]{barplot}} or
\code{\link[graphics]{boxplot}}.}
}
\description{
Plot variable importance
}
\examples{
# Single model
# vi_1 <- var_importance(preds$fitted_models$Model_ID_1)
# plot_importance(x = vi_1)

# Multiple models
# vi_2 <- var_importance(preds$fitted_models)
# plot_importance(x = vi_2)

}
