% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_glm.R
\name{predict_glm}
\alias{predict_glm}
\title{Model Prediction}
\usage{
predict_glm(model, newdata, clamping = FALSE, type = "response")
}
\arguments{
\item{model}{a \code{glm} fitted object.}

\item{newdata}{a data.frame or matrix with the new data to project the
predictions.}

\item{clamping}{(logical) whether to clamp values to a minimum and maximum
value, that are establish for the max and min values of the calibration
limits. Default = FALSE.}

\item{type}{(character) the type of prediction required. For a default
binomial model the default predictions are of log-odds (probabilities on
logit scale) and type = "response" gives the predicted probabilities.}
}
\value{
a \code{SpatRaster} object or a vector of probabilities.
}
\description{
Obtains predictions from a fitted generalized linear model object. It also
allows the clamping option to avoid extrapolation in areas outside the
calibration area.
}
