% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_selected.R
\name{predict_selected}
\alias{predict_selected}
\title{Predict the selected models}
\usage{
predict_selected(x, newdata, clamping = FALSE, type = "response")
}
\arguments{
\item{x}{the object returned by the function "calibration_glm()".}

\item{newdata}{a data.frame or matrix with the new data to project the
predictions.}

\item{clamping}{Clamp values to a minimum and maximum value, that are
establish for the max and min values of the calibration limits.}

\item{type}{the type of prediction required. For a default binomial model
the default predictions are of log-odds (probabilities on logit scale)
and type = "response" gives the predicted probabilities.}
}
\value{
a \code{SpatRaster} object or a list of vector with the probabilities.
}
\description{
Predict the selected models
}
