% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_selected.R
\name{predict_selected}
\alias{predict_selected}
\title{Predict the selected models}
\usage{
predict_selected(
  x,
  newdata,
  clamping = FALSE,
  type = "response",
  consensus = TRUE
)
}
\arguments{
\item{x}{the list object returned by the function
\code{\link{enmpa::calibration_glm}}.}

\item{newdata}{a \code{SpatRaster}, data.frame or matrix with the new data to
project the predictions.}

\item{clamping}{(logical) this option mitigates the risk of extreme
extrapolations when making model predictions for environmental conditions
beyond the calibrated data range. It employs the marginal values within the
calibration area to predict outcomes for more extreme conditions in transfer
areas. Default = FALSE.}

\item{type}{(character) the type of prediction required. For a default
binomial model the default predictions are of log-odds (probabilities on
logit scale) and type = "response" gives the predicted probabilities.}

\item{consensus}{(logical) whether to produce three consensus forecasts
(or “ensemble”) obtained by combining the forecasts from the collection the
selected models. Consensuses are calculated from the mean, median and weighted
mean using the wAIC as the weighting metric. An consensus map of variance
is also calculated to measure the amount of variability or disagreement among
individuals consensus. Default = TRUE.}
}
\value{
A list that includes a second list of fitting models, an individual model
predictions \code{SpatRaster} object, and a consensus predictions \code{SpatRaster}
object.
}
\description{
Wrapper function that facilitates the fitting, prediction of those models
selected as the most robust. In addition, it allows the calculation of
consensus models, when more than one model are selected.
}
