% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curve.R
\name{response_curve}
\alias{response_curve}
\title{Response curve}
\usage{
response_curve(model, variable, n = 100, new_data = NULL, new_range = NULL,
               rescale = FALSE)
}
\arguments{
\item{model}{an object of class \code{glm} or a list of them which inherit
from the class \code{lm}.}

\item{variable}{(character) vector, name or names of the variables to be
plotted.}

\item{n}{(numeric) an integer guiding the number of breaks. Default n = 100}

\item{new_data}{a \code{SpatRaster}, data.frame or  matrix of variables
representing the range of values for the complete extent of the study area.
Default = NULL.}

\item{new_range}{(numeric) a numerical vector with the lower and upper limits
of the variable. Default = NULL.}

\item{rescale}{(logical) whether to scale the y-axis to the probability
range (0-1). Default = FALSE.}
}
\description{
The function plot the response curve to  evaluates the variable of interest
and its limits.
}
