% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curve.R
\name{response_curve}
\alias{response_curve}
\title{Response curve}
\usage{
response_curve(model, variable, n = 100, new_data = NULL, extrapolate = FALSE,
               xlab = NULL, ylab = "Probability", col = "red", ...)
}
\arguments{
\item{model}{an object of class \code{glm} or a list of them which inherit
from the class \code{lm}.}

\item{variable}{(character) name of the variables to be plotted.}

\item{n}{(numeric) an integer guiding the number of breaks. Default n = 100}

\item{new_data}{a \code{SpatRaster}, data.frame or  matrix of variables
representing the range of values for the complete extent in an area of
interest. Default = NULL.}

\item{extrapolate}{(logical) whether to allow extrapolation to study the
behavior of the response outside the calibration limits. Ignored if \code{new_data}
is defined.}

\item{xlab}{(character) a label for the x axis.}

\item{ylab}{(character) a label for the y axis.}

\item{col}{(character) color for lines.}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}}.}
}
\description{
The function offers a comprehensive view of the response of a variable,
either based on a single model or multiple models. It plots the probabilities
of species presence across a wide range of values for a specific environmental
variable.

The function calculates these probabilities by focusing on a single
environmental gradient while keeping all other gradients constant at their
mean values (Elith et al., 2005). This approach allows for the estimation of
response curves, enabling a deeper understanding of how species occurrence
fluctuates in response to the environmental conditions captured by the
variable.
}
\examples{
# Load species occurrences and environmental data.
enm_data <- read.csv(system.file("extdata", "pa_data.csv", package = "enmpa"))
head(enm_data)

# Custom formulas
forms <- c("Sp ~ bio_1 + I(bio_1^2) + I(bio_12^2)",
           "Sp ~ bio_12 + I(bio_1^2) + I(bio_12^2)")

# Fitting models
fits <- fit_glm(forms, data = enm_data )

# Response curve for single models
response_curve(fits$Model_ID_1, variable = "bio_1",  extrapolate = TRUE)
response_curve(fits$Model_ID_1, variable = "bio_12", extrapolate = TRUE)

# Consensus response curve for both models
response_curve(fits, variable = "bio_1",  extrapolate = TRUE)
response_curve(fits, variable = "bio_12", extrapolate = TRUE)


}
