% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curve.R
\name{response_curve}
\alias{response_curve}
\title{Response curve: It evaluates the response of the variable and its limits}
\usage{
response_curve(model, variable, n = 100, new_data = NULL, new_range = NULL)
}
\arguments{
\item{model}{an object of class "glm" which inherits from the class "lm".}

\item{variable}{\code{character}, name of the variable to be plotted.}

\item{n}{\code{numeric}, an integer guiding the number of breaks. Default n = 100}

\item{new_data}{a \code{SpatRaster}, data.frame or  matrix of variables
representing the range of values for the complete extent of the study area.
Default = NULL}

\item{new_range}{\verb{numeric vector}, a numerical vector with the lower and
upper limits of the variable. Default = NULL}
}
\description{
Response curve: It evaluates the response of the variable and its limits
}
