% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sel_fit}
\alias{sel_fit}
\title{Example of selected models fitted}
\format{
A list with two elements.
\describe{
\item{ModelID_7}{a fitted glm.}
\item{selected}{a data.frame with results from evaluation of ModelID_7}
}
}
\usage{
sel_fit
}
\description{
A list containing fitted selected model(s) and the information
from model evaluation for such model(s).
}
\examples{
data("sel_fit", package = "enmpa")
}
\keyword{datasets}
