% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_importance.R
\name{var_importance}
\alias{var_importance}
\title{Variable importance}
\usage{
var_importance(x)
}
\arguments{
\item{x}{an object of class "glm" or a "list" of them which inherit
from the class "lm".}
}
\value{
\code{data.frame}
}
\description{
The function calculates the relative importance of the predictor variables
based on explained deviance.
}
